refactor: created OptionalPropertyEditor component and use it in RasterPropertyEditor
